EditArticleGUI semi-final version
Fully-functional insofar as GUI responsiveness and error messages. Just needs some debugging from Julio and Eyan related to the database and some database-related methods. Any other things that need to be known, either reference the comments in the program, or look at any info in the CreateArticleGUI push (since this is basically a copy and paste). Good work everyone!